[futures] Add c++ ivalue::Future collectAll() helper
Add some base c++ unittest coverage for ivalue::Future, and in
the process, add a basic collectAll() primitive, per 38937.

In the process, I realized that List<Future> is effectively
impossible to construct (since the Future's type is not templated,
but rather passed in,  the getTypePtr_<T>::call() isn't defined),
so added a workaround in List to make it possible.

Differential Revision: [D21756884](https://our.internmc.facebook.com/intern/diff/D21756884/)

ghstack-source-id: 104786677
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39119